Winbond developer pre-release #4
git-svn-id: https://subversion.assembla.com/svn/fakesmc@269 958d8502-8caa-4642-8a4d-47b05a98fa7a

+++ Plug-ins/FakeSMCSuperIO/SuperIO.h
@@ -118,7 +118,7 @@ protected:
 	
 	UInt16			m_Model;
 	
-	UInt8			m_RawVCore;
+	UInt16			m_RawVCore;
 	
 	bool			m_FanControl;
 	bool			m_FanVoltageControlled;
